+++ test/schema_statements/changing_column_type_test.rb
@@ -260,6 +260,30 @@ module SchemaStatements
       end
     end
 
+    def test_finalize_column_type_change_preserves_not_null_without_default_before_12
+      with_postgres(11) do
+        @connection.initialize_column_type_change(:projects, :description, :text)
+
+        assert_sql("UPDATE pg_catalog.pg_attribute") do
+          @connection.finalize_column_type_change(:projects, :description)
+        end
+
+        assert_equal false, column_for(:projects, :description).null
+      end
+    end
+
+    def test_finalize_column_type_change_preserves_not_null_without_default_after_12
+      with_postgres(12) do
+        @connection.initialize_column_type_change(:projects, :description, :text)
+
+        refute_sql("UPDATE pg_catalog.pg_attribute") do
+          @connection.finalize_column_type_change(:projects, :description)
+        end
+
+        assert_equal false, column_for(:projects, :description).null
+      end
+    end
+
     def test_finalize_column_type_change_keeps_columns_in_sync
       @connection.initialize_column_type_change(:projects, :id, :bigint)
       @connection.finalize_column_type_change(:projects, :id)