+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+## 0.8.2 (2023-09-26)
+
 - Promote check constraint to `NOT NULL` on PostgreSQL >= 12 when changing column type
 - Fix `safety_assured` with `revert`